refactor: extract from programBuilder test tools functions
