+++ __tests__/ProgramBuilder.test.tsx
@@ -1,6 +1,9 @@
 import { createMachine } from "xstate";
 import {
   fireEvent,
+  getAllTrainingSessionContainer,
+  getLastTrainingSessionContainer,
+  getTrainingSessionLastExercise,
   render,
   renderApp,
   waitFor,
@@ -8,107 +11,6 @@ import {
 } from "../tests/test.utils";
 import { createModel } from "@xstate/test";
 import invariant from "invariant";
-import { ReactTestInstance } from "react-test-renderer";
-
-function getAllTrainingSessionContainer({
-  screen,
-}: TestingContext): ReactTestInstance[] {
-  const trainingSessionContainerCollection = screen.queryAllByTestId(
-    /training-session-container.*/i
-  );
-
-  return trainingSessionContainerCollection;
-}
-
-/**
- * Throws error if no training session container found at all
- */
-function getLastTrainingSessionContainer(context: TestingContext): {
-  lastTrainingSessionId: string;
-  lastTrainingSessionContainer: ReactTestInstance;
-} {
-  const allTrainingSessionContainer = getAllTrainingSessionContainer(context);
-
-  invariant(
-    allTrainingSessionContainer.length > 0,
-    "Training session container not found"
-  );
-
-  const lastTrainingSessionContainer =
-    allTrainingSessionContainer[allTrainingSessionContainer.length - 1];
-
-  const lastTrainingSessionContainerTestId: string =
-    lastTrainingSessionContainer.props.testID;
-
-  const lastTrainingSessionIdSplit =
-    lastTrainingSessionContainerTestId.split("-");
-
-  const lastTrainingSessionId = lastTrainingSessionIdSplit
-    .slice(lastTrainingSessionIdSplit.length - 5)
-    .join("-");
-
-  invariant(
-    lastTrainingSessionId !== undefined,
-    "lastTrainingSessionId is undefined"
-  );
-
-  return {
-    lastTrainingSessionId,
-    lastTrainingSessionContainer,
-  };
-}
-
-function getAllTrainingSessionExercise({
-  context: { screen },
-  trainingSessionId,
-}: {
-  context: TestingContext;
-  trainingSessionId: string;
-}): ReactTestInstance[] {
-  const trainingSessionContainer = screen.getByTestId(
-    `training-session-container-${trainingSessionId}`
-  );
-
-  const allExerciseContainers = within(
-    trainingSessionContainer
-  ).queryAllByTestId(/training-session-exercise-container-.*/i);
-
-  return allExerciseContainers;
-}
-
-function getTrainingSessionLastExercise({
-  context,
-  trainingSessionId,
-}: {
-  context: TestingContext;
-  trainingSessionId: string;
-}): { lastExerciseContainer: ReactTestInstance; lastExerciseId: string } {
-  const allExercisesContainers = getAllTrainingSessionExercise({
-    context,
-    trainingSessionId,
-  });
-
-  invariant(
-    allExercisesContainers.length > 0,
-    "allExercisesContainers length is < to 0"
-  );
-
-  const lastExerciseContainer =
-    allExercisesContainers[allExercisesContainers.length - 1];
-
-  const lastExerciseTestId: string = lastExerciseContainer.props.testID;
-
-  const lastExerciseTestIdSplit = lastExerciseTestId.split("-");
-
-  const lastExerciseId = lastExerciseTestIdSplit
-    .slice(lastExerciseTestIdSplit.length - 5)
-    .join("-");
-
-  return {
-    lastExerciseContainer,
-    lastExerciseId,
-  };
-}
 
 test("User goes to program builder screen from home", async () => {
   const screen = renderApp();
@@ -146,7 +48,7 @@ const programBuilderTestMachine = createMachine({
 
               waitFor(() => {
                 const trainingSessionContainerCollection =
-                  getAllTrainingSessionContainer(context);
+                  getAllTrainingSessionContainer(context.screen);
 
                 expect(trainingSessionContainerCollection.length).toBe(
                   expectedTrainingSessionsCounter
@@ -171,7 +73,7 @@ const programBuilderTestMachine = createMachine({
               );
 
               const { lastTrainingSessionContainer } =
-                getLastTrainingSessionContainer(context);
+                getLastTrainingSessionContainer(context.screen);
 
               waitFor(() => {
                 const allRelatedExercisesContainer = within(
@@ -213,7 +115,7 @@ const programBuilderTestMachine = createMachine({
               );
 
               const { lastTrainingSessionContainer } =
-                getLastTrainingSessionContainer(context);
+                getLastTrainingSessionContainer(context.screen);
 
               waitFor(() => {
                 const allRelatedExercisesContainer = within(
@@ -266,7 +168,7 @@ const programBuilderTestMachine = createMachine({
           // Checking global expected counter
           waitFor(async () => {
             const trainingSessionContainerCollection =
-              getAllTrainingSessionContainer(context);
+              getAllTrainingSessionContainer(context.screen);
 
             expect(trainingSessionContainerCollection.length).toBe(
               expectedTrainingSessionsCounter
@@ -346,15 +248,16 @@ const programBuilderTestModel = createModel<TestingContext>(
 
       context.expectedTrainingSessionsCounter--;
 
-      const { lastTrainingSessionId } =
-        getLastTrainingSessionContainer(context);
+      const { lastTrainingSessionId } = getLastTrainingSessionContainer(
+        context.screen
+      );
 
       const removeLastTrainingSessionButton = await screen.findByTestId(
         `remove-training-session-button-${lastTrainingSessionId}`
       );
 
       context.lastlyRemovedTrainningSessionId = lastTrainingSessionId;
-      getAllTrainingSessionContainer(context);
+      getAllTrainingSessionContainer(context.screen);
       fireEvent.press(removeLastTrainingSessionButton);
     },
   },
@@ -368,7 +271,7 @@ const programBuilderTestModel = createModel<TestingContext>(
           : expectedTrainingSessionExerciseCounter + 1;
 
       const { lastTrainingSessionContainer, lastTrainingSessionId } =
-        getLastTrainingSessionContainer(context);
+        getLastTrainingSessionContainer(context.screen);
 
       const addExerciseButton = await within(
         lastTrainingSessionContainer
@@ -388,11 +291,12 @@ const programBuilderTestModel = createModel<TestingContext>(
       // @ts-ignore
       context.expectedTrainingSessionExerciseCounter--;
 
-      const { lastTrainingSessionId } =
-        getLastTrainingSessionContainer(context);
+      const { lastTrainingSessionId } = getLastTrainingSessionContainer(
+        context.screen
+      );
       const { lastExerciseContainer, lastExerciseId } =
         getTrainingSessionLastExercise({
-          context,
+          screen: context.screen,
           trainingSessionId: lastTrainingSessionId,
         });
